chore: move logtypes into logresponse
